feat: add login feature and a lot more (read long message)
so basically, I had to create a way to manage certain global values (token and user profile data). This was done by using Cubits. I also created an http interceptor and changed all services to work based on it, to sort of standarize the way we create services and prevent hard coding the reading of the Bearer token. Lastly, I coded the login feature. To do so, I had to instally the supabase dependency and ensure it initializes on the main app so the Auth service (which uses the supabase client) works as intended and we can both retrieve the authorization token and keep it updated even after closing the app safely. This is probably the largest commit I have ever done. May christ be with us.

+++ lib/features/favorites/presentation/bloc/favorites_bloc.dart
@@ -1,15 +1,12 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:tcompro_customer/features/favorites/data/favorite_service.dart';
-import 'package:tcompro_customer/features/favorites/domain/entities/favorite.dart';
 import 'package:tcompro_customer/features/favorites/presentation/bloc/favorites_event.dart';
 import 'package:tcompro_customer/features/favorites/presentation/bloc/favorites_state.dart';
 
 class FavoritesBloc extends Bloc<FavoritesEvent, FavoritesState> {
   final FavoriteService service;
-  final int customerId;
   
-  FavoritesBloc({required this.service, required this.customerId})
-      : super(FavoritesState.initial()) {
+  FavoritesBloc({required this.service}) : super(FavoritesState.initial()) {
     on<LoadFavoritesEvent>(_onLoadFavorites);
     on<ToggleFavoriteEvent>(_onToggleFavorite);
   }
@@ -17,34 +14,32 @@ class FavoritesBloc extends Bloc<FavoritesEvent, FavoritesState> {
   Future<void> _onLoadFavorites(
     LoadFavoritesEvent event, Emitter<FavoritesState> emit) async {
       try {
-        emit(state.copyWith(loading: true));
-        final favorites = await service.fetchFavorites(customerId);
+        emit(state.copyWith(
+          loading: true, 
+          customerId: event.customerId
+        ));
+
+        // Usamos el ID del evento para la petición
+        final favorites = await service.fetchFavorites(event.customerId);
+        
         emit(state.copyWith(favoriteProducts: favorites, loading: false));
       } catch (e) {
         emit(state.copyWith(loading: false));
       }
     }
 
-Future<void> _onToggleFavorite(
-    ToggleFavoriteEvent event, 
-    Emitter<FavoritesState> emit
-  ) async {
+  Future<void> _onToggleFavorite(ToggleFavoriteEvent event, Emitter<FavoritesState> emit) async {
+    final customerId = state.customerId;
+    if (customerId == null) return; 
+
     final isFavorite = state.favoriteProducts.any(
       (f) => f.productId == event.favorite.productId,
     );
 
-    final updated = List<Favorite>.from(state.favoriteProducts);
-
     if (isFavorite) {
-      updated.removeWhere((f) => f.productId == event.favorite.productId);
-      emit(state.copyWith(favoriteProducts: updated));
-
-      await service.removeFavorite(customerId, event.favorite.productId);
+       await service.removeFavorite(customerId, event.favorite.productId);
     } else {
-      updated.add(event.favorite);
-      emit(state.copyWith(favoriteProducts: updated));
-
-      await service.addFavorite(customerId, event.favorite.productId);
+       await service.addFavorite(customerId, event.favorite.productId);
     }
   }
 }
